apply suggestions for migrating display functions to utils folder
