+++ src/components/AlertList.js
@@ -1,26 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { useTheme } from '@material-ui/core/styles'
-import useMediaQuery from '@material-ui/core/useMediaQuery'
 import Skeleton from '@material-ui/lab/Skeleton'
 import GridList from '@material-ui/core/GridList'
 import GridListTile from '@material-ui/core/GridListTile'
+import { getGridListCols, useWidth } from 'utils/display'
 import Alert from './Alert'
-
-const getGridListCols = (width) => {
-  switch (width) {
-    case 'xl':
-      return 4
-    case 'lg':
-      return 4
-    case 'md':
-      return 3
-    case 'sm':
-      return 2
-    default:
-      return 1
-  }
-}
 
 function AlertSkeletonLoader({ limit }) {
   const rows = Array.from(Array(limit).keys())
@@ -42,18 +26,6 @@ AlertSkeletonLoader.propTypes = {
   limit: PropTypes.number.isRequired,
 }
 
-function useWidth() {
-  const theme = useTheme()
-  const keys = [...theme.breakpoints.keys].reverse()
-  return (
-    keys.reduce((output, key) => {
-      // eslint-disable-next-line react-hooks/rules-of-hooks
-      const matches = useMediaQuery(theme.breakpoints.up(key))
-      return !output && matches ? key : output
-    }, null) || 'xs'
-  )
-}
-
 function LoadAlertList({ data, width }) {
   if (!data || data.length === 0) {
     return (
@@ -65,7 +37,7 @@ function LoadAlertList({ data, width }) {
   }
 
   return (
-    <GridList cols={getGridListCols(width)}>
+    <GridList cols={getGridListCols[width]}>
       {data.map((item) => (
         <GridListTile key={item.AlertTitle} cols={1}>
           <Alert
